+++ src/main/java/org/summerboot/jexpress/boot/config/ConfigUtil.java
@@ -110,7 +110,7 @@ public class ConfigUtil {
         String timeoutDesc = BackOffice.agent.getProcessTimeoutAlertMessage();
 
         if (ConfigLoadMode.cli_format == mode) {
-            updated += formatConfig(Path.of("etc", "boot.conf").toFile().getAbsoluteFile(), BackOffice.agent, log);
+            updated += formatConfig(Path.of("etc", "boot.conf").toFile(), BackOffice.agent, log);
             for (String fileName : configs.keySet()) {
                 File configFile = Paths.get(configFolder.toString(), fileName).toFile();
                 try (var a = Timeout.watch("loading config file " + configFile, timeoutMs).withDesc(timeoutDesc)) {
@@ -155,7 +155,8 @@ public class ConfigUtil {
         }
     }
 
-    public static int formatConfig(File configFile, JExpressConfig cfg, Logger log) throws IOException {
+    public static int formatConfig(File targetFile, JExpressConfig cfg, Logger log) throws IOException {
+        File configFile = targetFile.getAbsoluteFile();
         ImportResource ir = (ImportResource) cfg.getClass().getAnnotation(ImportResource.class);
         if (ir != null && !ir.generateTemplate()) {
             return 0;